Prefixes iterator
